fix(android): startRouter 第二参 fd 转 Long
gomobile 把 Go 的 int 参数映射成 Java long,生成的签名是
startRouter(String, long);ParcelFileDescriptor.fd 是 Int,编译报
"Int but Long was expected"。fd.toLong() 修掉。

+++ android/app/src/main/kotlin/io/github/yiguihai11/smartproxy/SmartProxyVpnService.kt
@@ -98,8 +98,9 @@ class SmartProxyVpnService : VpnService() {
 
             val fd = builder.establish()?.fd ?: return false
             // Go engine AAR:gomobile bind(mobile 包)→ smartproxy.mobile.Mobile。
-            // 注意 gomobile 导出方法首字母小写(lowerFirst),StartRouter → startRouter。
-            smartproxy.mobile.Mobile.startRouter(configJson, fd)
+            // 注意 gomobile 导出方法首字母小写(lowerFirst),StartRouter → startRouter;
+            // Go 的 int 参数在 Java 侧是 long,fd(Int)要转 Long。
+            smartproxy.mobile.Mobile.startRouter(configJson, fd.toLong())
             true
         } catch (e: Exception) {
             // config 解析 / Builder 参数 / establish / StartRouter 任一失败都走这。
